PMM-5804 Updating UI tests job

+++ pmm/pmm2-ui-tests.groovy
@@ -3,7 +3,7 @@ library changelog: false, identifier: 'lib@master', retriever: modernSCM([
     remote: 'https://github.com/Percona-Lab/jenkins-pipelines.git'
 ]) _
 
-void runStaging(String DOCKER_VERSION, CLIENT_VERSION, CLIENTS, CLIENT_INSTANCE, SERVER_IP) {
+void runStagingServer(String DOCKER_VERSION, CLIENT_VERSION, CLIENTS, CLIENT_INSTANCE, SERVER_IP) {
     stagingJob = build job: 'aws-staging-start', parameters: [
         string(name: 'DOCKER_VERSION', value: DOCKER_VERSION),
         string(name: 'CLIENT_VERSION', value: CLIENT_VERSION),
@@ -16,6 +16,7 @@ void runStaging(String DOCKER_VERSION, CLIENT_VERSION, CLIENTS, CLIENT_INSTANCE,
         string(name: 'DAYS', value: '1')
     ]
     env.VM_IP = stagingJob.buildVariables.IP
+    env.SERVER_IP = env.VM_IP
     env.VM_NAME = stagingJob.buildVariables.VM_NAME
     def clientInstance = "yes";
     if ( CLIENT_INSTANCE == clientInstance ) {
@@ -28,6 +29,32 @@ void runStaging(String DOCKER_VERSION, CLIENT_VERSION, CLIENTS, CLIENT_INSTANCE,
         env.PMM_UI_URL = "http://${VM_IP}"
     }
 }
+
+void runStagingClient(String DOCKER_VERSION, CLIENT_VERSION, CLIENTS, CLIENT_INSTANCE, SERVER_IP) {
+    stagingJob = build job: 'aws-staging-start', parameters: [
+        string(name: 'DOCKER_VERSION', value: DOCKER_VERSION),
+        string(name: 'CLIENT_VERSION', value: CLIENT_VERSION),
+        string(name: 'CLIENTS', value: CLIENTS),
+        string(name: 'CLIENT_INSTANCE', value: CLIENT_INSTANCE),
+        string(name: 'SERVER_IP', value: SERVER_IP),
+        string(name: 'NOTIFY', value: 'false'),
+        string(name: 'DAYS', value: '1')
+    ]
+    env.VM_CLIENT_IP = stagingJob.buildVariables.IP
+    env.VM_CLIENT_NAME = stagingJob.buildVariables.VM_NAME
+    env.VM_IP = stagingJob.buildVariables.SERVER_IP
+    def clientInstance = "yes";
+    if ( CLIENT_INSTANCE == clientInstance ) {
+        env.PMM_URL = "http://admin:admin@${SERVER_IP}"
+        env.PMM_UI_URL = "http://${SERVER_IP}"
+    }
+    else
+    {
+        env.PMM_URL = "http://admin:admin@${VM_IP}"
+        env.PMM_UI_URL = "http://${VM_IP}"
+    }
+}
+
 
 
 void destroyStaging(IP) {
@@ -132,51 +159,63 @@ pipeline {
             }
         }
         stage('Checkout Commit') {
-             when {
+            when {
                 expression { env.GIT_COMMIT_HASH.length()>0 }
             }
             steps {
                 sh 'git checkout ' + env.GIT_COMMIT_HASH
             }
         }
-        stage('Start staging') {
+        stage('Start PMM Server') {
+            when {
+                expression { env.CLIENT_INSTANCE == "no" }
+            }
             steps {
-                runStaging(DOCKER_VERSION, CLIENT_VERSION, '--addclient=ps,1 --addclient=mo,2 --with-replica --addclient=pgsql,1 --addclient=pxc,1 --with-proxysql --pmm2 --setup-alertmanager --add-annotation', CLIENT_INSTANCE, SERVER_IP)
+                runStagingServer(DOCKER_VERSION, CLIENT_VERSION, '', CLIENT_INSTANCE, SERVER_IP)
             }
         }
-        stage('Sanity check') {
+        stage('Start PMM Client Instance') {
             steps {
-                sh 'timeout 100 bash -c \'while [[ "$(curl -s -o /dev/null -w \'\'%{http_code}\'\' \${PMM_URL}/ping)" != "200" ]]; do sleep 5; done\' || false'
+                runStagingClient(DOCKER_VERSION, CLIENT_VERSION, '--addclient=ps,1 --addclient=mo,2 --with-replica --addclient=pgsql,1 --addclient=pxc,1 --with-proxysql --pmm2 --setup-alertmanager --add-annotation --setup-replication-ps-pmm2', 'yes', env.SERVER_IP)
             }
         }
-        stage('Setup Node') {
-            steps {
-                sh """
-                    curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.34.0/install.sh | bash
-                    . ~/.nvm/nvm.sh
-                    nvm install 12.14.1
-                    sudo rm -f /usr/bin/node
-                    sudo ln -s ~/.nvm/versions/node/v12.14.1/bin/node /usr/bin/node
-                    pushd pmm-app/
-                    npm install
-                    node -v
-                    npm -v
-                    sudo yum install -y gettext
-                    envsubst < env.list > env.generated.list
-                    popd
-                """
-            }
-        }
-        stage('Sleep') {
-            steps {
-                sh """
-                curl --data '{"enable_stt": true, "enable_telemetry": true}' -u admin:admin -X POST ${PMM_UI_URL}/v1/Settings/Change
-                curl -u admin:admin -X POST ${PMM_UI_URL}/v1/management/SecurityChecks/Start
-                """
-                sleep 300
-                sh """
-                curl --data '{"disable_stt": true, "enable_telemetry": true}' -u admin:admin -X POST ${PMM_UI_URL}/v1/Settings/Change
-                """
+        stage('Setup') {
+            parallel {
+                stage('Sanity check') {
+                    steps {
+                        sh 'timeout 100 bash -c \'while [[ "$(curl -s -o /dev/null -w \'\'%{http_code}\'\' \${PMM_URL}/ping)" != "200" ]]; do sleep 5; done\' || false'
+                    }
+                }
+                stage('Setup Node') {
+                    steps {
+                        sh """
+                            curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.34.0/install.sh | bash
+                            . ~/.nvm/nvm.sh
+                            nvm install 12.14.1
+                            sudo rm -f /usr/bin/node
+                            sudo ln -s ~/.nvm/versions/node/v12.14.1/bin/node /usr/bin/node
+                            pushd pmm-app/
+                            npm install
+                            node -v
+                            npm -v
+                            sudo yum install -y gettext
+                            envsubst < env.list > env.generated.list
+                            popd
+                        """
+                    }
+                }
+                stage('Sleep') {
+                    steps {
+                        sh """
+                        curl --data '{"enable_stt": true, "enable_telemetry": true}' -u admin:admin -X POST ${PMM_UI_URL}/v1/Settings/Change
+                        curl -u admin:admin -X POST ${PMM_UI_URL}/v1/management/SecurityChecks/Start
+                        """
+                        sleep 300
+                        sh """
+                        curl --data '{"disable_stt": true, "enable_telemetry": true}' -u admin:admin -X POST ${PMM_UI_URL}/v1/Settings/Change
+                        """
+                    }
+                }
             }
         }
         stage('Run AMI Setup & UI Tests') {
@@ -227,7 +266,16 @@ pipeline {
                 ./pmm-app/node_modules/.bin/mochawesome-merge pmm-app/tests/output/parallel_chunk*/*.json > pmm-app/tests/output/combine_results.json
                 ./pmm-app/node_modules/.bin/marge pmm-app/tests/output/combine_results.json --reportDir pmm-app/tests/output/ --inline --cdn --charts
             '''
-            destroyStaging(VM_NAME)
+            script {
+                if(env.VM_NAME)
+                {
+                    destroyStaging(VM_NAME)
+                }
+                if(env.VM_CLIENT_NAME)
+                {
+                    destroyStaging(VM_CLIENT_NAME)
+                }
+            }
             uploadAllureArtifacts()
             script {
                 if (currentBuild.result == null || currentBuild.result == 'SUCCESS') {
